requirements.txt
django==3.1.3
python-decouple==3.3 
# pillow==8.0.1
djangorestframework==3.12.2
djangorestframework_simplejwt==4.5.0 
django-cors-headers==3.5.0
arrow==0.17.0 
# psycopg2==2.8.6
WeasyPrint==52.1
pdfkit==0.6.1
requests==2.22.0
django-versatileimagefield==2.0
ipstack==0.1.4
geopy==2.0.0
pyfcm==1.4.9

